+++ QiLuTreeSeedlings/UI/UserCenterViewController/UserCenterViews/MySupplyViewController/ZIKMySupplyViewController.m
@@ -47,14 +47,14 @@
 
 - (void)requestData {
     [self requestSupplyRestrict];
-    [self requestSellList:[NSString stringWithFormat:@"%ld",self.page]];
+    [self requestSellList:[NSString stringWithFormat:@"%ld",(long)self.page]];
     __weak typeof(self) weakSelf = self;//解决循环引用的问题
     [self.mySupplyTableView addHeaderWithCallback:^{
-        [weakSelf requestSellList:[NSString stringWithFormat:@"%ld",weakSelf.page]];
+        [weakSelf requestSellList:[NSString stringWithFormat:@"%ld",(long)weakSelf.page]];
     }];
     [self.mySupplyTableView addFooterWithCallback:^{
         weakSelf.page++;
-        [weakSelf requestSellList:[NSString stringWithFormat:@"%ld",weakSelf.page]];
+        [weakSelf requestSellList:[NSString stringWithFormat:@"%ld",(long)weakSelf.page]];
     }];
 
 }
